pyproject.toml
[project]
name = "bobc-stablecoin"
version = "0.1.0"
description = "Overcollateralized algorithmic stablecoin pegged to the BOB backed by Ether. Made with Vyper 🐍"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hypothesis>=6.122.1",
    "pytest>=8.3.4",
    "ruff>=0.8.1",
    "snekmate>=0.1.0",
    "titanoboa>=0.2.5",
    "vyper>=0.4.0",
]
